requirements.txt
asgiref==3.8.1
Django==5.1.1
django-computedfields==0.2.6
django-fast-update==0.2.3
sqlparse==0.5.1
typing_extensions==4.12.2
